+++ module/modules/account/dal/user.go
@@ -168,6 +168,10 @@ func QueryUserByID(helper dbhelper.DBHelper, id int) (model.UserDetail, bool) {
 
 // DeleteUser 删除用户，根据用户ID
 func DeleteUser(helper dbhelper.DBHelper, id int) bool {
+	if id <= 0 {
+		return false
+	}
+
 	sql := fmt.Sprintf("delete from account_user where id =%d", id)
 	_, ret := helper.Execute(sql)
 	return ret
@@ -175,6 +179,10 @@ func DeleteUser(helper dbhelper.DBHelper, id int) bool {
 
 // DeleteUserByAccount 删除用户，根据用户账号&密码
 func DeleteUserByAccount(helper dbhelper.DBHelper, account, password string) bool {
+	if account == "admin" {
+		return false
+	}
+
 	sql := fmt.Sprintf("delete from account_user where account ='%s' and password='%s'", account, password)
 	_, ret := helper.Execute(sql)
 	return ret
